requirements.txt
# Django REST server
django==2.1.7

# Timezone
pytz==2018.9

# Configuration
# https://django-environ.readthedocs.io/en/latest/
django-environ==0.4.3

# Python-PostgreSQL Database Adapter
psycopg2==2.7.7
psycopg2-binary==2.7.7

# Django model mixins and utilities
django-model-utils==3.1.2